Fixes to subscription
