+++ app/services/create/create_subscription.rb
@@ -1,8 +1,8 @@
 class CreateSubscription
-  def call(current_user, email, subscription_params)
-    subscription = current_user ? current_user.subscription.build(subscription_params) : Subscription.new(subscription_params)
+  def call(current_user, subscription_params)
+    subscription = current_user ? current_user.build_subscription(subscription_params) : Subscription.new(subscription_params)
     if subscription.valid?
-      EmailDispatchesWorker.perform_async(email)
+      EmailDispatchesWorker.perform_async(subscription_params[:email])
     end
     subscription
   end
