+++ spec/controllers/web/subscriptions_controller_spec.rb
@@ -15,7 +15,7 @@ RSpec.describe Web::SubscriptionsController, type: :controller do
     end
 
     context 'with valid attributes and without user' do
-      subject { post 'create', params: { subscription: attributes_for(:subscription, :with_no_user) }}
+      subject { post 'create', params: { subscription: attributes_for(:guest_subscription) }}
 
       it 'saves the new job to database' do
         expect { subject }.to change(Subscription.all, :count).by(1)
@@ -23,6 +23,16 @@ RSpec.describe Web::SubscriptionsController, type: :controller do
 
       it 'redirects to root view' do
         expect(subject).to redirect_to(root_path)
+      end
+    end
+
+    context 'with current_user' do
+      let(:user) { create(:user, :junior) }
+      subject { post 'create', params: { subscription: attributes_for(:guest_subscription) }}
+      before { login_user(user) }
+
+      it 'saves the new subscription to database' do
+        expect { subject }.to change(Subscription.all, :count).by(1)
       end
     end
   end
