+++ spec/factories/subscriptions.rb
@@ -4,14 +4,10 @@ FactoryBot.define do
     active true
 
     trait :with_user do
-      user_id 1
-    end
-    
-    trait :with_no_user do
-      user_id nil
+      association :user, factory: :user, strategy: :build
     end
 
     factory :user_subscription,  traits: [:with_user]
-    factory :guest_subscription, traits: [:with_no_user]
+    factory :guest_subscription
   end
 end
