+++ spec/rails_helper.rb
@@ -11,6 +11,7 @@ require 'support/factory_bot'
 require 'support/shoulda_matchers'
 require 'support/sorcery'
 require 'support/active_job'
+require 'sidekiq/testing'
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
 # spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are
